requirements.txt
easydict==1.9
imageio==2.14.0
matplotlib==3.5.1
numpy==1.22.0
opencv_python_headless==4.5.5.62
pillow==9.0.1
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
setuptools==58.0.4
six==1.16.0
sklearn
tensorboardx==2.4.1
torch==1.8.0
torchvision==0.9.0
tqdm==4.62.3